Move utility script
